Fix bin/setup command for ppl w/o required bundler
Bundler 1.17.3 is a specific requirement so we install it when its not
there at setup otherwise it results in something like:

---
Traceback (most recent call last):
	2: from /Users/khalil/.rbenv/versions/2.7.2/bin/bundle:23:in `<main>'
	1: from /Users/khalil/.rbenv/versions/2.7.2/lib/ruby/2.7.0/rubygems.rb:296:in `activate_bin_path'
/Users/khalil/.rbenv/versions/2.7.2/lib/ruby/2.7.0/rubygems.rb:277:in `find_spec_for_exe': Could not find 'bundler' (1.17.3) required by your /Users/khalil/code/Private/kelder/Gemfile.lock. (Gem::GemNotFoundException)
To update to the latest version installed on your system, run `bundle update --bundler`.
To install the missing version, run `gem install bundler:1.17.3`
---

Which is not that nice.
(and because I'm bashing a bit more lately 😅)